#35 refactor : InputBasic 컴포넌트 분리
