+++ components/input/InputBasic/MainInput.tsx
@@ -1,81 +1,32 @@
 import { ModalTwoButton } from 'components/common/Modal';
-import Toast from 'components/common/Toast';
-import { MODAL_DELETE_ALL, MODAL_DELETE_SCREENSHOT, MODAL_LEAVE_PAGE } from 'constants/modal';
+import { MODAL_DELETE_SCREENSHOT, MODAL_LEAVE_PAGE } from 'constants/modal';
+import { workInputState } from 'core/atom';
 import useRouteChangeBlocking from 'lib/hooks/input/useRouteChangeBlocking';
 import useModal from 'lib/hooks/useModal';
-import {
-  IcBtnDeleteFile,
-  IcBtnScreenshot,
-  IcBtnScreenshotHide,
-  IcFileCheckbox,
-  IcFileCheckboxAfter,
-  IcKatchupLogo,
-} from 'public/assets/icons';
-import { useEffect, useRef, useState } from 'react';
+import { IcBtnScreenshot, IcBtnScreenshotHide } from 'public/assets/icons';
+import { useEffect, useState } from 'react';
+import { useRecoilState } from 'recoil';
 
 import styled from '@emotion/styled';
 
 import CardModal from '../InputCard/CardModal';
 import { ScreenshotInput } from '../InputScreenshot';
+import FileInput from './FileInput';
+import WorkInput from './WorkInput';
 
 const MainInput = () => {
-  const [workInput, setWorkInput] = useState('');
-  const [letterCount, setLetterCount] = useState(0);
-
-  const [isChecked, setIsChecked] = useState(false);
-
-  const [fileInput, setFileInput] = useState<File[]>([]);
-  const fileInputRef = useRef<HTMLInputElement>(null);
-
+  const [workInput, setWorkInput] = useRecoilState(workInputState);
   const [isScreenshotShowing, setIsScreenshotShowing] = useState(false);
 
-  const [toastMessage, setToastMessage] = useState('');
-  const [toastKey, setToastKey] = useState<number>();
-
   const cardModal = useModal();
-  const deleteAllModal = useModal();
   const leavePageModal = useModal();
   const screenshotCancelModal = useModal();
+
   const { offRouteChangeBlocking } = useRouteChangeBlocking(leavePageModal.toggle, workInput);
 
   const handleScreenshotShowing = (e: React.MouseEvent<HTMLButtonElement>) => {
-    setFileInput([]);
     setIsScreenshotShowing((prev) => !prev);
     screenshotCancelModal.toggle();
-  };
-
-  const handleInputChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
-    const inputText = e.target.value;
-    const inputCount = inputText.length;
-
-    setWorkInput(inputText);
-    setLetterCount(inputCount);
-  };
-
-  const handleDeleteAll = (e: React.MouseEvent<HTMLButtonElement>) => {
-    setWorkInput('');
-    setLetterCount(0);
-    deleteAllModal.toggle();
-  };
-
-  const handleFileSelect = (e: React.ChangeEvent<HTMLInputElement>) => {
-    const files = e.target.files;
-    if (files && files.length > 0) {
-      const file = files[0];
-      setFileInput((prev) => [...prev, file]);
-    }
-  };
-
-  const handleFileBtnClick = () => {
-    fileInputRef.current?.click();
-  };
-
-  const handleDeleteFile = (file: File) => {
-    setFileInput((prev) => prev.filter((selectedFile) => selectedFile !== file));
-  };
-
-  const handleCheckboxChange = () => {
-    setIsChecked((prev) => !prev);
   };
 
   const handleNext = (e: React.MouseEvent<HTMLButtonElement>) => {
@@ -94,78 +45,17 @@ const MainInput = () => {
     };
   }, []);
 
-  useEffect(() => {
-    if (isChecked) {
-      setToastMessage('곧 추가될 예정이에요! Coming soon..');
-      setToastKey(Date.now());
-    }
-  }, [isChecked]);
-
   return (
     <StMainInputWrapper>
       <StMainInput>
-        <StDeleteAllBtn type="button" onClick={deleteAllModal.toggle}>
-          모든 내용 지우기
-        </StDeleteAllBtn>
-        <StInputWrapper>
-          <h2>
-            <IcKatchupLogo /> 업무 내용
-          </h2>
-          <StWorkInput>
-            <textarea
-              maxLength={2000}
-              value={workInput}
-              onChange={handleInputChange}
-              placeholder="업무 내용을 입력해주세요"
-            />
-            <p>
-              <span>{letterCount}</span>/2000자
-            </p>
-          </StWorkInput>
-        </StInputWrapper>
-        <StFileWrapper>
-          <StFileSelect>
-            <StFileBtnWrapper>
-              <h2>
-                <IcKatchupLogo /> 파일 첨부
-              </h2>
-              <input ref={fileInputRef} type="file" onChange={handleFileSelect} />
-              <button onClick={handleFileBtnClick}>파일선택</button>
-            </StFileBtnWrapper>
-            <label>
-              <input type="checkbox" checked={isChecked} onChange={handleCheckboxChange} />
-              {isChecked ? <IcFileCheckboxAfter /> : <IcFileCheckbox />}
-              파일명 자동 변경
-            </label>
-          </StFileSelect>
-          <StFileInput>
-            {fileInput.length ? (
-              fileInput.map((file, index) => (
-                <p key={index}>
-                  <button onClick={() => handleDeleteFile(file)}>
-                    <IcBtnDeleteFile />
-                  </button>
-                  {file.name}
-                  <span>{`${(file.size / (1024 * 1024)).toFixed(2)}MB`}</span>
-                </p>
-              ))
-            ) : (
-              <StEmpty>
-                <p>파일을 선택하거나 드래그하세요</p>
-                <p>파일당 최대 크기: 10MB</p>
-              </StEmpty>
-            )}
-          </StFileInput>
-        </StFileWrapper>
+        <WorkInput />
+        <FileInput />
         <StNextBtn disabled={!workInput.length}>
           <button type="button" onClick={cardModal.toggle}>
             다음 단계
           </button>
         </StNextBtn>
         <CardModal isShowing={cardModal.isShowing} handleHide={cardModal.toggle} />
-        <StToastWrapper>
-          <Toast key={toastKey} message={toastMessage} />
-        </StToastWrapper>
       </StMainInput>
       {isScreenshotShowing ? (
         <>
@@ -186,14 +76,6 @@ const MainInput = () => {
         rightButtonName={'벗어나기'}
         handleLeftButton={leavePageModal.toggle}
         handleRightButton={() => offRouteChangeBlocking()}
-      />
-      <ModalTwoButton
-        isShowing={deleteAllModal.isShowing}
-        contents={MODAL_DELETE_ALL}
-        leftButtonName={'취소하기'}
-        rightButtonName={'내용 지우기'}
-        handleLeftButton={deleteAllModal.toggle}
-        handleRightButton={handleDeleteAll}
       />
       <ModalTwoButton
         isShowing={screenshotCancelModal.isShowing}
@@ -236,198 +118,6 @@ const StMainInput = styled.section`
   background-color: ${({ theme }) => theme.colors.katchup_white};
 
   box-shadow: 0px 0px 20px 0px rgba(0, 0, 0, 0.05);
-`;
-
-const StToastWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-
-  width: 100%;
-`;
-
-const StDeleteAllBtn = styled.button`
-  position: absolute;
-  top: 5rem;
-  right: 7.5rem;
-
-  border: none;
-  background: transparent;
-  color: ${({ theme }) => theme.colors.katchup_gray};
-  ${({ theme }) => theme.fonts.p2_text};
-  text-decoration: underline;
-`;
-
-const StInputWrapper = styled.div`
-  padding: 6.4rem 7.4rem 3.4rem 7.4rem;
-
-  & > h2 {
-    display: flex;
-    align-items: center;
-    gap: 0.8rem;
-
-    margin-bottom: 1.4rem;
-
-    ${({ theme }) => theme.fonts.h1_title};
-  }
-`;
-
-const StWorkInput = styled.div`
-  position: relative;
-
-  width: 100%;
-  height: 45.8rem;
-  padding: 4rem;
-
-  border: 0.1rem solid ${({ theme }) => theme.colors.katchup_line_gray};
-  border-radius: 0.8rem;
-
-  & > textarea {
-    width: 100%;
-    height: 32.7rem;
-
-    border: none;
-    outline: none;
-    resize: none;
-    ${({ theme }) => theme.fonts.p1_text};
-
-    scrollbar-width: 0.5rem;
-    scrollbar-color: ${({ theme }) => theme.colors.katchup_gray};
-
-    ::placeholder {
-      color: ${({ theme }) => theme.colors.katchup_gray};
-    }
-
-    ::-webkit-scrollbar {
-      width: 0.5rem;
-    }
-
-    ::-webkit-scrollbar-thumb {
-      background-color: ${({ theme }) => theme.colors.katchup_gray};
-      border-radius: 3rem;
-    }
-  }
-
-  & > p {
-    position: absolute;
-    bottom: 2rem;
-    right: 4rem;
-
-    color: ${({ theme }) => theme.colors.katchup_dark_gray};
-    ${({ theme }) => theme.fonts.caption};
-
-    & > span {
-      color: ${({ theme }) => theme.colors.katchup_main};
-      ${({ theme }) => theme.fonts.caption};
-    }
-  }
-`;
-
-const StFileWrapper = styled.div`
-  padding: 0rem 7.4rem;
-`;
-
-const StEmpty = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-
-  height: 100%;
-
-  color: ${({ theme }) => theme.colors.katchup_gray};
-
-  & > p {
-    ${({ theme }) => theme.fonts.p1_text};
-  }
-
-  & > p:last-child {
-    ${({ theme }) => theme.fonts.p3_text};
-  }
-`;
-
-const StFileSelect = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  margin-bottom: 1.4rem;
-
-  input {
-    display: none;
-  }
-
-  & > label {
-    display: flex;
-    align-items: center;
-
-    color: ${({ theme }) => theme.colors.katchup_black};
-    ${({ theme }) => theme.fonts.p3_text};
-  }
-`;
-
-const StFileBtnWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  & > h2 {
-    display: flex;
-    align-items: center;
-    gap: 0.8rem;
-
-    margin-right: 1.4rem;
-
-    ${({ theme }) => theme.fonts.h1_title};
-  }
-  & > button {
-    padding: 0.5rem 1.5rem;
-
-    border: 0.1rem solid ${({ theme }) => theme.colors.katchup_line_gray};
-    border-radius: 0.8rem;
-    background-color: ${({ theme }) => theme.colors.katchup_white};
-    ${({ theme }) => theme.fonts.p3_text};
-  }
-`;
-
-const StFileInput = styled.div`
-  position: relative;
-  overflow-y: scroll;
-
-  width: 100%;
-  height: 11.4rem;
-  padding: 1.2rem 2.2rem;
-
-  border: 0.1rem solid ${({ theme }) => theme.colors.katchup_line_gray};
-  border-radius: 0.8rem;
-
-  & > p {
-    display: flex;
-    align-items: center;
-    gap: 1.6rem;
-
-    ${({ theme }) => theme.fonts.p1_text};
-  }
-
-  & > p > button {
-    padding: 0;
-
-    border: none;
-    background: transparent;
-  }
-
-  & > p > span {
-    color: ${({ theme }) => theme.colors.katchup_dark_gray};
-    ${({ theme }) => theme.fonts.caption};
-  }
-
-  ::-webkit-scrollbar {
-    width: 0.5rem;
-  }
-
-  ::-webkit-scrollbar-thumb {
-    background-color: ${({ theme }) => theme.colors.katchup_gray};
-    border-radius: 3rem;
-  }
 `;
 
 const StNextBtn = styled.div<{ disabled: boolean }>`
