+++ components/input/InputBasic/index.ts
@@ -1,1 +1,3 @@
 export { default as MainInput } from './MainInput';
+export { default as FileInput } from './FileInput';
+export { default as WorkInput } from './WorkInput';
